remove old layout components
