+++ src/App.vue
@@ -9,24 +9,14 @@
 <script setup lang="ts">
 import { Layouts } from "@/layouts/types";
 import { computed, defineAsyncComponent } from "vue";
-import {
-	injectStrict,
-	AUTH_MODULE_KEY,
-	ENV_CONFIG_MODULE_KEY,
-} from "@/utils/inject";
+import { injectStrict, AUTH_MODULE_KEY } from "@/utils/inject";
 import { useRoute } from "vue-router";
 
 const authModule = injectStrict(AUTH_MODULE_KEY);
-const envConfigModule = injectStrict(ENV_CONFIG_MODULE_KEY);
 const route = useRoute();
 
-const shouldUseNewLayout = envConfigModule.getEnv.FEATURE_NEW_LAYOUT_ENABLED;
-const defaultLayout = shouldUseNewLayout
-	? Layouts.NEW_LOGGED_IN
-	: Layouts.LOGGED_IN;
-
 const layout = computed(() => {
-	let layout = (route.meta.layout as Layouts) || defaultLayout;
+	let layout = (route.meta.layout as Layouts) || Layouts.LOGGED_IN;
 
 	if (!authModule.isLoggedIn) {
 		layout = Layouts.LOGGED_OUT;
