--- src/components/README.md
@@ -1,7 +1,0 @@
-# Components
-
-## Directory Structure
-
-Where should my new component be located?
-
-[![Atomic Design - How to categorize](https://res.cloudinary.com/practicaldev/image/fetch/s--IHiwlC8g--/c_limit%2Cf_auto%2Cfl_progressive%2Cq_auto%2Cw_880/https://thepracticaldev.s3.amazonaws.com/i/4gd0sforrkoi0g8a4gdw.png)](https://dev.to/maciekchmura/how-i-structure-a-react-project-3c2i)
