--- src/layouts/README.md
@@ -1,3 +1,0 @@
-# LAYOUTS
-
-This directory contains your Application Layouts.
