+++ src/layouts/types.ts
@@ -1,6 +1,5 @@
 export enum Layouts {
-	LOGGED_IN = "loggedIn",
+	LOGGED_IN = "LoggedIn",
 	LOGGED_OUT = "loggedOut",
 	LERN_STORE = "lernStore",
-	NEW_LOGGED_IN = "newLoggedIn",
 }
